requirements.txt
Flask==2.3.2
flask_cors==4.0.0
numpy==1.26.0
psycopg2==2.9.6
python-dotenv==1.0.0
python_nmap==0.7.1
Requests==2.31.0
setproctitle==1.3.2
six==1.16.0
